refactor: 수정사항 수정, 리팩토링
