+++ src/main/java/com/techeer/cokkiri/domain/example/domain/entity/User.java
@@ -1,18 +1,18 @@
 package com.techeer.cokkiri.domain.example.domain.entity;
 
-import com.techeer.cokkiri.global.domain.BaseEntity;
+import com.techeer.cokkiri.global.entity.BaseEntity;
 import java.util.ArrayList;
 import javax.persistence.*;
 import lombok.*;
 
 @Getter
 @Entity
-@AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Table (name = "USERS")
 public class User extends BaseEntity {
   @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  @Column(name = "USER_ID")
+  @GeneratedValue(strategy = GenerationType.SEQUENCE)
+  @Column(name = "user_id")
   private Long id;
 
   // 회원 아이디
